Refactor: AccountDeletionNotice 컴포넌트 분리
